CR by Robbie Hoodless addressed bar unit tests and comments
