+++ src/main/test/TestRequestManagement.java
@@ -4,7 +4,6 @@ import requestManagement.Dispatcher;
 import requestManagement.RequestManager;
 import requestManagement.Service;
 import requestManagement.fleetManager.FleetManager;
-import requestManagement.fleetManager.FleetManagerImplementation;
 import requestManagement.loadBalancing.LoadBalancer;
 
 import java.util.ArrayList;
@@ -61,7 +60,7 @@ class TestRequestManagement {
     void testRequestManagerBuilder() {
         Dispatcher dispatcher = new Dispatcher(services);
         LoadBalancer loadBalancer = new LoadBalancer();
-        FleetManager fleetManager = FleetManagerImplementation.getInstance();
+        FleetManager fleetManager = FleetManager.getInstance();
 
         /* Test that the request manager builds successfully with valid member variables */
         RequestManager requestManager = RequestManager.getBuilder()
